models seperated and functioning
